chore: Update build process and dependencies
This commit updates the build process and dependencies for the skycrane project. It adds a step to install the latest nightly version of Rust and includes the rustfmt and clippy components. Additionally, it fetches the wasm plugin spec from a remote location and sets the SKYCRANE_SPEC_PATH environment variable. The Makefile has been modified to include a new build-dev target and the cargo build command has been updated to use cargo build instead of make build. These changes improve the build process and ensure that the project is using the latest tools and dependencies.

Signed-off-by: Victor Palade <victor@cloudflavor.io>